==> 4-northeast-analysis.Rmd <==
---
title: "4-northeast-analysis"
author: "Daniel Hocking"
date: "August 28, 2014"
output: html_document
---

```{r}

rm(list=ls())
gc()

```


```{r results = 'hide', warning=FALSE, message=FALSE}
library(ggplot2)
library(ggmcmc)
library(dplyr)
#library(nlme)
library(devtools)
install_github("Conte-Ecology/conteStreamTemperature")
library(conteStreamTemperature)

baseDir <- getwd() # works as long as you have the project loaded in RStudio - does not work for kniting

dataInDir <- paste0(baseDir, '/dataIn/')
dataOutDir <- paste0(baseDir, '/dataOut/')
dataLocalDir <- paste0(baseDir, '/localData/')
graphsDir <- paste0(baseDir, '/graphs/')

#source(paste0(baseDir, 'code/functions/temperatureModelingFunctions.R'))

# load standardized and formatted data created with the 3-statModelPrep.Rmd script
load(paste0(dataOutDir, 'tempDataSync.RData'))
```

```{r eval = FALSE}
print(getwd())
```

```{r site within HUC model with residual AR1}

######
# rename all variables before setting the fixed and random, then select them by list?
####

tempDataSyncS$landcover0 <- ifelse(tempDataSyncS$landcover == 0, 1, 0)
tempDataSyncS$landcover1 <- ifelse(tempDataSyncS$landcover == 1, 1, 0)
tempDataSyncS$landcover5 <- ifelse(tempDataSyncS$landcover == 5, 1, 0)
tempDataSyncS$landcover6 <- ifelse(tempDataSyncS$landcover == 6, 1, 0)
tempDataSyncS$landcover7 <- ifelse(tempDataSyncS$landcover == 7, 1, 0)
tempDataSyncS$landcover8 <- ifelse(tempDataSyncS$landcover == 8, 1, 0)
tempDataSyncS$landcover9 <- ifelse(tempDataSyncS$landcover == 9, 1, 0)
tempDataSyncS$landcover10 <- ifelse(tempDataSyncS$landcover == 10, 1, 0)
tempDataSyncS$landcover11 <- ifelse(tempDataSyncS$landcover == 11, 1, 0)

 fixed.ef <- c("intercept" 
                        , "Lat" 
                        , "Long"
               , "landcover1"
               , "landcover5"
               , "landcover6"
               , "landcover7"
               , "landcover8"
               , "landcover9"
               , "landcover10"
               , "landcover11"
               , "flow_accum"
                        , "slope" 
                        , "Lake" 
                        , "glacier" 
               , "airTemp.flow"
                        )

site.ef <- c( "intercept.site" 
                      , "airTemp" 
                      #, "airTempLagged2" 
                      , "prcp" 
                      #, "prcpLagged1" 
                      , "airTemp.prcp"
                      )

year.ef <- c( "intercept.year"
                      , "dOY" 
                      , "dOY2"
                      , "dOY3"
                      )

cov.list <- list(fixed.ef = fixed.ef, site.ef = site.ef, year.ef = year.ef)
# model matrix not working because creates a design matrix

#data.cal <- prepDF(tempDataSyncS, formulae = formulae)
#str(data.cal)

monitor.params <- c(#"residuals",
            #"deviance",
            "sigma",
            "B.ar1",
            "mu.ar1",
            "sigma.ar1",
            "B.0",
            "B.site",
            "rho.B.site",
            "mu.site",
            "sigma.b.site",
            "B.huc",
            "rho.B.huc",
            "mu.huc",
            "sigma.b.huc",
            "B.year",
            "rho.B.year",
            "mu.year",
            "sigma.b.year")

coda.tf <- T # currently only works in full for TRUE (using coda.samples)
system.time(M.ar1 <- modelRegionalTempAR1(tempDataSyncS, cov.list = cov.list, firstObsRows = firstObsRows, evalRows = evalRows, n.burn = 1000, n.it = 1000, n.thin = 1, nc = 3, coda = coda.tf, param.list = monitor.params)) # Slow with AR1: ~3-6 min per 100 iterations (13 min per 100 iter for site AR)

save(M.ar1, file = paste0(dataLocalDir, "mcmc.RData"))
save(cov.list, file = paste0(dataLocalDir, "covariate-list.RData"))

```

## Evaluate MCMC Iterations

```{r check model convergence and mixing}
system.time(ggs.ar1 <- ggs(M.ar1)) # slow and memory intensive - 2 minutes with residuals saved in mcmc object and no other programs running but uses most of the memory on my 8 GB laptop
gc() # clears all of the temporary variables stored by ggs and cuts hanging memory more than in half.

ggmcmc(ggs.ar1, file = 'localData/ggmcmc-B0.pdf', family = "B.0", plot = c("ggs_traceplot", "ggs_compare_partial", "ggs_autocorrelation"))

ggmcmc(ggs.ar1, file = 'localData/ggmcmc-mu-huc.pdf', family = "mu.huc", plot = c("ggs_traceplot", "ggs_compare_partial", "ggs_autocorrelation"))

ggmcmc(ggs.ar1, file = 'localData/ggmcmc-mu-year.pdf', family = "mu.year", plot = c("ggs_traceplot", "ggs_compare_partial", "ggs_autocorrelation"))

ggmcmc(ggs.ar1, file = 'localData/ggmcmc-mu-ar1.pdf', family = "mu.ar1", plot = c("ggs_traceplot", "ggs_compare_partial", "ggs_autocorrelation"))

ggmcmc(ggs.ar1, file = 'localData/ggmcmc-sigma-ar1.pdf', family = "sigma.ar1", plot = c("ggs_traceplot", "ggs_compare_partial", "ggs_autocorrelation"))

ggmcmc(ggs.ar1, file = 'localData/ggmcmc-sigma-site.pdf', family = "sigma.b.site", plot = c("ggs_traceplot", "ggs_compare_partial", "ggs_autocorrelation"))

ggmcmc(ggs.ar1, file = 'localData/ggmcmc-sigma-huc.pdf', family = "sigma.b.huc", plot = c("ggs_traceplot", "ggs_compare_partial", "ggs_autocorrelation"))

ggmcmc(ggs.ar1, file = 'localData/ggmcmc-sigma-year.pdf', family = "sigma.b.year", plot = c("ggs_traceplot", "ggs_compare_partial", "ggs_autocorrelation"))

ggmcmc(ggs.ar1, file = 'localData/ggmcmc-ar1-rho-huc.pdf', family = "rho.B.huc", plot = "ggs_traceplot")

ggmcmc(ggs.ar1, file = 'localData/ggmcmc-ar1-B-ar1.pdf', family = "B.ar1", plot = c("ggs_traceplot", "ggs_compare_partial", "ggs_autocorrelation"))

gc()
```

## Check model fit

```{r check model fit}
#------ Check residual patterns------

system.time(coef.summary <- avgCoefs(ggs.ar1)) # 15s
tempDataSyncS$resid.ar1 <- dplyr::filter(coef.summary, grepl('^residuals', coef.summary$Parameter))$mean

# Observed vs Residuals
g <- ggplot(data.frame(tempDataSyncS), aes(temp, resid.ar1)) 
#g + geom_point(alpha = 0.3) + geom_smooth() + theme_bw()

library(hexbin)
g + stat_binhex(bins = 100) + geom_smooth() + theme_bw()

g + geom_point() + geom_density2d() + geom_smooth() + theme_bw()

# Residuals by random HUC8
b <- ggplot(tempDataSyncS, aes(x = HUC8, y = resid.ar1))
b + geom_boxplot() + coord_flip()

# Residuals by random site within HUC8
b <- ggplot(tempDataSyncS, aes(x = site, y = resid.ar1))
b + geom_boxplot() + coord_flip()

# Residuals by random year
ggplot(tempDataSyncS, aes(x = as.factor(year), y = resid.ar1)) + geom_boxplot() + coord_flip()

# Residuals by covariates
ggplot(tempDataSyncS, aes(Forest, resid.ar1)) + geom_point(alpha = 0.5) + geom_smooth() + theme_bw()

ggplot(tempDataSyncS, aes(airTemp, resid.ar1)) + geom_point(alpha = 0.5) + geom_smooth() + theme_bw()

ggplot(tempDataSyncS, aes(dOY, resid.ar1)) + geom_point(alpha = 0.5) + geom_smooth() + theme_bw()

ggplot(tempDataSyncS, aes(dOY^2, resid.ar1)) + geom_point(alpha = 0.5) + geom_smooth() + theme_bw()

# Check Autocorrelation of residuals
# correlation in MCMC
acf(ar1.mat[ , "residuals[1]"], lag.max = 40, xlab = "Lag", ylab = "Correlation", main = "")
acf(ar1.mat[ , "residuals[2]"], lag.max = 40, xlab = "Lag", ylab = "Correlation", main = "")

# correlation in residuals over time 
acf(tempDataSyncS$resid.ar1, lag.max = 100, xlab = "Lag", ylab = "Correlation", main = "") # not perfect but better than for non-AR model

```


```{r test}

gc()

library(rjags)



B.fixed <- dplyr::filter(coef.summary, grepl('^B.0', coef.summary$Parameter))
B.fixed$coef <- cov.list$fixed.ef

mu.huc <- nameCoefs(coef.summary = coef.summary, rand.levels = levels(as.factor(tempDataSyncS$HUC8)), family = "mu.huc", name = "coef", conditional = FALSE, coefs = cov.list$site.ef)

mu.year <- nameCoefs(coef.summary = coef.summary, rand.levels = levels(as.factor(tempDataSyncS$year)), family = "mu.year", name = "coef", conditional = FALSE, coefs = cov.list$year.ef)

mu.ar1 <- dplyr::filter(coef.summary, grepl('^mu.ar1', coef.summary$Parameter))
mu.ar1$coef <- mu.ar1$Parameter

sigma.b.site <- nameCoefs(coef.summary = coef.summary, rand.levels = levels(as.factor(tempDataSyncS$site)), family = "sigma.b.site", conditional = FALSE, coefs = cov.list$site.ef, name = "coef")

sigma.b.huc <- nameCoefs(coef.summary = coef.summary, rand.levels = levels(as.factor(tempDataSyncS$site)), family = "sigma.b.huc", conditional = FALSE, coefs = cov.list$site.ef, name = "coef")

sigma.b.year <- nameCoefs(coef.summary = coef.summary, rand.levels = levels(as.factor(tempDataSyncS$year)), family = "sigma.b.year", conditional = FALSE, coefs = cov.list$year.ef, name = "coef")

sigma.ar1 <- dplyr::filter(coef.summary, grepl('^sigma.ar1', coef.summary$Parameter))

B.ar1 <- nameCoefs(coef.summary = coef.summary, rand.levels = levels(as.factor(tempDataSyncS$site)), family = 'B.ar1', name = "site")

B.site <- nameCoefs(coef.summary = coef.summary, rand.levels = levels(as.factor(tempDataSyncS$site)), family = "B.site", coefs = cov.list$site.ef, name = "site")

B.huc <- nameCoefs(coef.summary = coef.summary, rand.levels = levels(as.factor(tempDataSyncS$HUC8)), family = "B.huc", coefs = cov.list$site.ef, name = "huc")

B.year <- nameCoefs(coef.summary = coef.summary, rand.levels = levels(as.factor(tempDataSyncS$year)), family = "B.year", coefs = cov.list$year.ef, name = "year")


# Make correlation matrix of random huc effects
cor.huc <- as.data.frame(matrix(NA, length(cov.list$site.ef), length(cov.list$site.ef)))
names(cor.huc) <- cov.list$site.ef
row.names(cor.huc) <- cov.list$site.ef
for(k in 1:length(cov.list$site.ef)){
  for(k.prime in 1:length(cov.list$site.ef)){
    cor.huc[k, k.prime] <- coef.summary[which(coef.summary$Parameter == paste('rho.B.huc[',k,',',k.prime,']', sep="")), "mean"]
  }
}
cor.huc <- round(cor.huc, digits=3)
cor.huc[upper.tri(cor.huc, diag=TRUE)] <- ''
cor.huc

# Make correlation matrix of random year effects
cor.year <- as.data.frame(matrix(NA, length(cov.list$year.ef), length(cov.list$year.ef)))
names(cor.year) <- cov.list$year.ef
row.names(cor.year) <- cov.list$year.ef
for(l in 1:length(cov.list$year.ef)){
  for(l.prime in 1:length(cov.list$year.ef)){
    cor.year[l, l.prime] <- coef.summary[which(coef.summary$Parameter == paste('rho.B.year[',l,',',l.prime,']', sep="")), "mean"]
  }
}
cor.year <- round(cor.year, digits=3)
cor.year[upper.tri(cor.year, diag=TRUE)] <- ''
cor.year



(fix.ef <- rbind(B.fixed, select(mu.huc, -index), select(mu.year, -index), mu.ar1))

coef.list <- list(B.fixed = B.fixed
                  , mu.huc = mu.huc
                  , mu.year = mu.year
                  , mu.ar1 = mu.ar1
, sigma.site = sigma.b.site
, sigma.huc = sigma.b.huc
, sigma.year = sigma.b.year
, sigma.ar1 = sigma.ar1
, cor.huc = cor.huc
, cor.year = cor.year
, B.site = B.site
, B.huc = B.huc
, B.year = B.year
, B.ar1 = B.ar1
)

save(coef.list, file = "dataOut/coef-MA.Rdata")

```


# old using S4 object

# combine model summary results into an S4 Object
setClass("jagsSummary",
         representation(fixEf="data.frame",
                        ranEf="list",
                        ranCor="list",
                        BSite="data.frame",
                        BHUC="data.frame",
                        BYear="data.frame"))

modSummary <- new("jagsSummary")
modSummary@fixEf <- fix.ef
modSummary@ranEf <- list(ranSite=ran.ef.site, ranHUC=ran.ef.huc, ranYear=ran.ef.year)
modSummary@ranCor <- list(corSite=cor.huc, corYear=cor.year)
modSummary@BSite <- B.site
modSummary@BHUC <- B.huc
modSummary@BYear <- B.year

# modSummary <- NULL
# modSummary$fixEf <- fix.ef
# modSummary$ranEf <- list(ranSite=ran.ef.site, ranYear=ran.ef.year)
# modSummary$ranCor <- list(corSite=cor.site, corYear=cor.year)
# modSummary$BSite <- B.site
# modSummary$BYear <- B.year

#modSummary
str(modSummary)
head(modSummary@fixEf)
head(modSummary@ranEf)
head(modSummary@ranCor)
head(modSummary@BSite)
head(modSummary@BHUC)
head(modSummary@BYear)




# predict to new sites
test.df <- data.frame(matrix(NA, 4, length(names(tempDataSyncS))))
names(test.df) <- names(tempDataSyncS)
test.df$site <- c("one", "two", "three", "MAUSGS_WEST_BROOK")
test.df$HUC8 <- c("a", "b", "c", "01070004")

data.val <- prepDF(test.df)
str(data.val)

test.B.site <- test.df %>%
  dplyr::filter(!(site %in% levels(as.factor(tempDataSync$site)))) %>%
  dplyr::select(site) %>%
  dplyr::mutate(mean = 0)

B.site.new <- rbind(B.site[ , c("site", "mean")], test.B.site)




```{r predict}
tempDataSyncS$tempPredicted <- NA
tempDataSyncS$tempPredicted <- predictTemp(data = tempDataSyncS, coefs = coef.list, firstObsRows = firstObsRows, evalRows = evalRows)

tempDataSyncValidS$tempPredicted <- NA
tempDataSyncValidS$tempPredicted <- predictTemp(data = tempDataSyncValidS, firstObsRows = firstObsRows.valid, evalRows = evalRows.valid, coefs = coef.list)


library(ggplot2)
ggplot(tempDataSyncValidS, aes(temp, tempPredicted)) + geom_point() + geom_abline(aes(1,1), colour = 'blue')
tempDataSyncValidS$resid.r <- tempDataSyncValidS$temp - tempDataSyncValidS$tempPredicted
rmse(tempDataSyncValidS$resid.r)


library(ggplot2)
ggplot(tempDataSyncS, aes(temp, tempPredicted)) + geom_point() + geom_abline(aes(1,1), colour = 'blue')

tempDataSyncS$resid.r <- tempDataSyncS$temp - tempDataSyncS$tempPredicted

rmse(tempDataSyncS$resid.r)
rmse(tempDataSyncS$resid.ar1)

tempDataSyncS$tempPredicted.jags <- tempDataSyncS$temp - tempDataSyncS$resid.ar1

#summary(tempDataSyncS)

ggplot(tempDataSyncS, aes(tempPredicted.jags, tempPredicted)) + geom_point()

```


```{r ending info}

print(sessionInfo())

```
